+++ internal/amd64/bytes_sse2_amd64.go
@@ -34,6 +34,9 @@ func maskBitsSSE2(dst []byte, b []byte, c byte)
 func maskBitsLessSSE2(dst []byte, b []byte, c byte)
 
 //go:noescape
+func maskBitsAny4SSE2(dst []byte, b []byte, chars uint32)
+
+//go:noescape
 func maskBitsAnySSE2(dst []byte, b []byte, chars uint64)
 
 //go:noescape